# internal/zstd: optimize window reset using make for buffer allocation